requirements.txt
PyQt6>=6.4.0
PyMuPDF>=1.23.0
Pillow>=9.3.0 